MDEV-21117 Incremental commit fixes a use case
of ineffective (does not change data) "blank" transaction that does
not create any engine branch but still gets binlogged when STATEMENT
binlog format.

Fixed with refining the binlogging and recovery parts.
At binlogging the fact of no engine involvement is passed to Gtid ctor
through a new THD::is_1pc_ro_trans. The new member is meaningful only
for ha_commit_one_phase() "real" transaction.
For replication the blank status of transaction is encoded as -1 or extra engines.
This only applies to FL_TRANSACTIONAL group.

At recovery the blank group is identified to be treated
as safe to truncate for the semisync slave recovery.