MDEV-23510: arm64 lf_hash
volatile has no memory barrier schemantics, its for mmaped IO
so lets allow some optimizer gains and stop pretending it helps
with memory atomicity.

The MDEV lists a SEGV an assumption is made that an address was
partially read. As C packs structs strictly and on arm64 the
cache line size is 128, move the 32bit hashnr to the end so
we don't get the *key pointer split across two cache lines.